DeBugging

+++ dao/src/main/java/hu/unideb/cardcommunity/api/CardDao.java
@@ -7,7 +7,7 @@ import hu.unideb.cardcommunity.model.Card;
 public interface CardDao {
 	Card save(Card entity);
 	List<Card> save(List<Card> entities);
-	Card findById(int id);
+	Card findById(long id);
 	Card findByName(String name);
 	List<Card> findAll();
 }
