+++ dao/src/main/java/hu/unideb/cardcommunity/api/CardMapperDao.java
@@ -7,6 +7,6 @@ import hu.unideb.cardcommunity.model.CardMapper;
 public interface CardMapperDao {
 	CardMapper save(CardMapper entity);
 	List<CardMapper> save(List<CardMapper> entities);
-	CardMapper findById(int id);
+	CardMapper findById(long id);
 	List<CardMapper> findAll();
 }
