+++ dao/src/main/java/hu/unideb/cardcommunity/api/DeckDao.java
@@ -7,7 +7,7 @@ import hu.unideb.cardcommunity.model.Deck;
 public interface DeckDao {
 	Deck save(Deck entity);
 	List<Deck> save(List<Deck> entities);
-	List<Deck> findByName(String name, int userId);
-	List<Deck> findAll(int gameTypeId,int userId);
-	List<Deck> findByUserId(int userId);
+	List<Deck> findByName(String name, long userId);
+	List<Deck> findAll(long gameTypeId,long userId);
+	List<Deck> findByUserId(long userId);
 }
