+++ dao/src/main/java/hu/unideb/cardcommunity/api/RuleDeckDao.java
@@ -7,6 +7,6 @@ import hu.unideb.cardcommunity.model.RuleDeck;
 public interface RuleDeckDao {
 	RuleDeck save(RuleDeck entity);
 	List<RuleDeck> save(List<RuleDeck> entities);
-	RuleDeck findById(int id);
-	List<RuleDeck> findByDeckId(int id);
+	RuleDeck findById(long id);
+	List<RuleDeck> findByDeckId(long id);
 }
