+++ dao/src/main/java/hu/unideb/cardcommunity/api/UserAccountDao.java
@@ -7,7 +7,7 @@ import hu.unideb.cardcommunity.model.UserAccount;
 public interface UserAccountDao {
 	UserAccount save(UserAccount entity);
 	List<UserAccount> save(List<UserAccount> entities);
-	UserAccount findById(int id);
+	UserAccount findById(long id);
 	List<UserAccount> findByName(String username);
 	List<UserAccount> findAll();
 }
